pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Oleg Cherkashin <impuls_64@mail.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.4"
gunicorn = "^20.1.0"
django-bootstrap4 = "^22.3"
whitenoise = "^6.2.0"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
dj-database-url = "^1.2.0"
django-filter = "^22.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
